Add ImPlot to Project, and update Window Defauts
I added the Library ImPlot to the project, and was able to
successfully Update a few boolean values in the ImGui and ImPlot
demo.cpp files respectively, so that the program starts with those
settings enabled. Shows the ImPlot default demo window, with the
ImGui default demo window showing the Docking example.